+++ App.tsx
@@ -13,7 +13,7 @@ import dayjs from 'dayjs';
 import {AuthProvider} from './src/contexts/AuthContext';
 import {Settings} from 'react-native-fbsdk-next';
 import VersionCheck from 'react-native-version-check';
-import DeviceInfo from 'react-native-device-info';
+import storeVersion from 'react-native-store-version';
 dayjs.extend(buddhaEra);
 import {
   firebaseInitialize,
@@ -29,7 +29,6 @@ import {PointProvider} from './src/contexts/PointContext';
 import moment from 'moment';
 import RNExitApp from 'react-native-kill-app';
 
-import packageJson from './package.json';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 moment.updateLocale('th', {
   relativeTime: {
@@ -47,7 +46,7 @@ const ActionContextState = {
 };
 
 const ActionContext = createContext<ActionContextType>(ActionContextState);
-const latestVersion = packageJson.version;
+
 const App = () => {
   const [actiontaskId, setActiontaskId] = useState<string | null>('');
   const requestTracking = async () => {
@@ -66,19 +65,24 @@ const App = () => {
   const checkVersion = async () => {
     const isIOS = Platform.OS === 'ios';
     const currentVersion = VersionCheck.getCurrentVersion();
-
-    const needUpdate = await VersionCheck.needUpdate({
-      currentVersion,
-      latestVersion,
-    });
-    const packageName = DeviceInfo.getBundleId();
     const storeUrl = await VersionCheck.getAppStoreUrl({
       appID: '6443516628',
-      appName: 'เรียกโดรน - ไอคอนเกษตร',
     });
 
     const playStoreUrl = await VersionCheck.getPlayStoreUrl({
-      packageName: 'com.iconkaset.droner',
+      packageName: 'com.iconkaset.farmer',
+    });
+
+    const {remote} = await storeVersion({
+      version: currentVersion,
+      androidStoreURL: playStoreUrl,
+      iosStoreURL: storeUrl,
+      country: 'TH',
+    });
+
+    const needUpdate = await VersionCheck.needUpdate({
+      currentVersion,
+      latestVersion: remote,
     });
 
     if (needUpdate.isNeeded) {